Improved logging, error returns for startup functions, title/company data
All startup functions now return an error object which is checked
and logged appropiatly. A bugfix was applied to the loggin library
as its verbosity check for standard out was messed up.

The header template now takes the site name and company name which
have been added to the configuration.