MAC handling, finished registration page, logout button
All MAC handling in the application is done by passing net.HardwareAddr
objects instead of just strings. By converting them as soon as they
enter the application it provides better consistancy and error handling
as the net.ParseMAC() function handles more formats then the old regex.
A special case is added where a user enters a MAC with no colons,
hyphens, or periods. In which case if the string is a valid 12 chars
long then it's formatted into the 3 group dotted notation.

The registration page is finished. With some "fancy" handling it will
do both manual and automatic registrations with the same page. The main
difference is whether or not a text input is displayed to enter a MAC
address. The JS intellegently get's the value if the element exists.

The registration page will not show a login form if the user is already
logged in except in the case where the user's machine is not registered.
It will then show a login page regardless of logged in status.

A logout function has been added.

With this commit the user facing application is finished (as far as I
can tell). The remaining pieces are the admin area and an integrated
DHCP server that interfaces with the database thereby removing the
need to manage a host and leases file.