Created admin area, currently search and view any user
Search can be on an IP, MAC, or username. Usernames may be partial
and will use a wildcard on the right side. Currently IP search
is not working as other pieces are needed for that.

Also removed code to parse the dhcpd leases file. A DHCP server
will be embedded so will not need to parse files. The write hosts
file system was also removed.