Moved routes back to single file, admin dash and manage users
It's a lot easier to manage the routes when they're all in a single
file. So the controllers now expose their route handlers and the routes
are defined in the server package's routes.go file.

The admin dashboard is back. Admins and help desk users can see it and
look at other user's devices. Admin actions aren't working yet.
Nor is admin search. Nor is user management. It's coming. It's close.
Maybe tomorrow. I'm tired.