Support multiple databases, log message for helpdesk user and db connect
Created a Database section in the configuration file to specify
a DB type, address, username, and password. For now it still only
supports SQLite with the database file in the address field. Eventually
this can be expanded to other DB types.

Fixed SQLite schema file to delete a table if it exists.

Show log message to specify when a helpdesk user logs in. Also show
a log message to show the DB type and address being used.