Manage dependencies with godep again
It was a mistake to try and manage them on my own. Godep has been
around longer and does a better job at copying the needed files.

However it has a bug where it copies dot-files like .travis.yml
and .gitignore. These files aren't necessary and shouldn't be copied.

If the deps are updated, run `make clean_vendor` to remove them.