Fixed Windows issue, lease storage, more log info, fixed device api
Windows clients don't like when you send a 16 bytes server identifier
option. Linux apparently doesn't care. That was a bug the dhcp4
dependency. I've forked the repo and fixed the problem hince the
change in dependency path. Fixed a bug in the device api where the
formUser for an auto registration wasn't being put in the correct
scope. It was stuck in the "if" scope causing a panic further down
when the code tried to access its members.

Added more logging information to the DHCP server, it logs every request
and any relevent data about the request as it goes through its motions.

Leases are now stored in the database upon ACK and RELEASE. If any other
part of the application wishes to get lease information it should get
it from the database. The database will act as the authoritative source
of information for the rest of the application. The DHCP server will
keep a working copy of lease information in memory and will save new
information as needed.