Added panic recovery, fixed bug with lease renewal
The DHCPACK message was returning the request IP option
instead of the lease IP. When a client renews an exising
IP it doesn't send that option meaning the DHCPACK was returning
a 0.0.0.0 address.

Added a panic recovery function in ServeDHCP so like the HTTP
server, it one request panics it doesn't bring down the whole
application.