Added cache control to requests
The way netreg worked, all URLs would resolve to the
registration server. This will make it so only requests
to the real domain name will cache. Others will not cache.

This should solve any issues of people going back to
their original request only to see a ghost of the registration
page.