Logs path, api simplified, integrated DHCP, clean up
Logs directory setting was changed to path as both a directory
or a file may be used. A file would be more appropiate in prod.

For the API, /api/device/register and /api/device/delete are both
now /api/device. The HTTP method determines the action. POST means
register device, DELETE means, well, delete.

The DHCP server is now integrated in the main application. It is
off by default and must be explicitly enabled in the config. The
old DHCP binary that was mainly used for testing is now cleaned
up and can be used to run the DHCP system as a separate application.
It uses the same configuration file as the main application along
with it own specific file.

Cleaned up code in the webserver startup methods and the new
logger methods.