Removed auth code from devices api, panic catcher
The devices api controller no longer handles its own auth.
The frontend makes a separate call to /login before making
the registration call if the user needs to login. In the
process I also discovered a bug that caused automatic regs
to panic if a machine attempts to register when it hasn't
received a lease from the DHCP server. I corrected the bug
and added a panic catcher middleware at the top of the middleware
chain to return a 500 HTTP code and log the panic error.

Also stopped logging when bcrypt errors for different hashes.
Obviously it will have different hashes, that's the point. No
need to log it.

Added convienence function to create an empty API response object.