Dist recipe, divide by 0, test configs, version text
Created a distribution recipe in make. It will compile the
application and tar it with the configs, templates, and other
things it needs.

Fixed a bug where a new install will panic at the admin dashboard
due to dividing by 0. We can't have the universe explode.

Added flags to binary to test configurations. -test will test
the main configuration. -testd will test the DHCP configuration.
They cannot be tested at the same time.

-version will display version information.