Added description field to devices
The description field can be used to organize a user's devices
however they see fit. Many times the platforms are not expressive
enought and the user agents can just be confusing. Later on some
sort of filtering system can used to filter unwanted words. However
only admins and the user themselves will see the description so
it may not be a high priority.