Added job scheduler, added "rolling" device expiration mode
The job scheduler is responsible for cleaning up the database
periodically by deleting expired devices and users.

A new device expiration mode is available: "rolling". If a device
is set to never (value of 0) expire it will never be collected by
the job scheduler. If a device is set to "rolling" (value of 1)
it will not be expired exactly, but it will be subject to the
device collector job.

Use never for truly permanent devices, use rolling for normal devices.