Fixed styling bug in header
With the new logo-wrapper, the title was being pushed
below the header bar. Also it was being underlined like
a normal link when below the mobile width threshold.

This update makes the header take up as much height as it needs,
make the wrapper a static 100px high, and collapses at 900px
instead of 550 to account for longer titles. If you have a title
that goes past the 900 mark, make it shorter

[ci skip]