Fixes bug where client switches network but gets the same address
This commit fixes a bug where if a client switched from one network
to another and attempted a dhcp REQUEST, it would receive its old
IP address instead of a new one on the new network.

This fix adds a check to see if the Gateway IP is non-zero
in which case it will use the gateway IP to the network and
not the client requested IP.

Also reduced the time to wait for a ping.
Increased lease offer time.