Separated DHCP into separate application
Tests showed that the DHCP server was simply too busy and clogged
up the rest of the system. In particular, although it explicitaly
yields to Go scheduler, the task scheduler ran very irregularly.
To alleviate this, the DHCP server will run as a separate service.
The main application already "communicates" via the database so it's
easy enough to separate. Plus the DHCP package was basically isolated
anyway. In the process just about everything has been changed to private
including types and functions. Documentation and service files will
follow to orchistrate the two applications.