Fixed bugs in (un)install scripts, set address default to empty
There were several bugs in the install/uninstall scripts in particular
detecting the presence of Systemd and AppArmor. There was also a bug
where the Systemd service wasn't being removed on uninstall.

The webserver.address configuration setting is switched back to
empty as default. It will be recommanded to change it to localhost.
But for first setup and testing it's easier to have it listen on
all addresses.