Implemented Guest registration
Currently verification tokens only show in the logs.
The next step is to provide multiple different means of sending
verification codes to users. The first two will be email an Twilio
for SMS.

A user model is never saved for a guest. All devices registered for
a guest will have a description that starts with "Guest -" followed
by their name.