Separate upgrade script
To reduce accidental install instead of upgrade and to have a clean
separation, the upgrade script is now separate from install. Also,
an upgrade will now require the user to manually extract the new
dist tarball to the correct place and run the upgrade script. This
will make it so an upgrade doesn't rely on an older version of the
upgrade script.