Implemented variable permission system
We needed a third level of permissions so it was high time to fix
permissions once and for all. It's implemented to bitwise operations
on unsigned 64-bit ints each bit being a particular permission. There
are currently 3 predefined permission sets or groups. Long term this
could be implemented in such a way as to be configurable by the end
user. For now it makes it a lot easier to add a new permission level
if needed.

All calls to IsAdmin, IsHelpDesk, and IsNormal have been replaced
with a call to Can() and checking for a particular permission vs
group membership.

Other changes in this commit:

- Bug fix for displaying Daily @ time on admin user edit page
- Added Copyright section to contributing guidelines
- Added device last_seen column to device manage page
- Created new readonly user (pervious help desk user) and gave the
help desk user more permissions
- Formatted validity dates on user list page
- Removed Onesimus Systems copyright statements and transitioned to
University of Southern Indiana
- Permission checks have been delegated to the handlers instead of
middleware due to the myriad of permissions

References #6