Fix bug where response was broadcasted when using a Gateway
Some Windows 7 clients did not properly handle a unicast response
from the server. When it toggled to broadcast bit = 1, the broadcast
packets weren't being delivered correctly. In some environments they
were. But in more complex setups the packets were never making it.
Now the server will always unicast back to a relay if giaddr is set
and will otherwise honor the broadcast bit if I ever add the
feature to use DHCP on a local subnet.