Added external upgrade script to upgrade from new tarball
The current upgrade script is specific to a particular version
of PG. Each upgrade may be slightly different and so it's a bad
idea to rely on an old version to upgrade a new one. The new
pg-update script will move the current PG to a temp old directory,
extract the new version and run the new version's upgrade.sh.

pg-upgrade will only exist after an install so it is safe to always
assume an upgrade. The uninstall will remove the script. The script
will be in /usr/local/bin which should be in the $PATH of most
Linux systems.