Run pg-upgrade from /tmp, cp backup of uninstall script, run as root
If any of the scripts are started by non-root use, they will reexecute
themselves using sudo.

pg-upgrade had a bug due to the way bash executes scripts. When the
upgrade script overwrites pg-upgrade, the execution fails near the end.
Instead it will copy itself to a temp directory and exec from there.

During install and upgrade, the uninstall script will be copied to
the data directory so it can be used to purge the configuration
in case the user forgot to use the -p flag.