Added reports package, created blacklist reports
The reports package provides and easy way to generate reports in the
admin UI. A report is nothing more than a function that is registered
with a short name and long name. The long name is show to the user
and the short name is used as the URL slug. Reports will typically be
paired with a template to display the data to the user. The templates
may generate links as appropiate to assist the user.

The first reports are for the blacklist. One shows blacklisted users
and the other blacklisted devices.