Fixed context issues with latest version of Gorilla mux/context
Gorilla mux was using the new context package in Go 1.7 which
is incompatible with the use of context outside the mux package.
The environment, session, and user where being assigned to one Request
pointer which was then changed inside mux as it was manipulating the
context. All context functions have been separated in to build tagged
files so they will use Gorilla context for less than Go 1.7 and
the native context package for Go 1.7+.