Issue #16 - Added MySQL support, not fully tested
To facilitate easier addition of new databases and to provide some
customization, the different database setup functions have been
separated into different files. Build tags can be used to build
support for all the databases (default with the makefile), or for
a sepecific database. The tags are dbmysql, dbsqlite, dbpostgres, and
dball. The tags are fairly self-explainitory. The makefile uses dball.
The main benefit of building for only a subset of databases is to
disable CGO. The MySQL and eventual Postgres packages are pure Go
unlike the SQLite package. So, but compiling without dbsqlite, CGO
can be turned off. This can provide much smaller Docker images
and make it easier to build on non-Linux platforms.