Don't set the sql_mode, let the client do it
If the sql_mode doesn't have ANSI and TRADITIONAL, PG will error
on start. The admin will need to edit the my.cnf file and set the
mode, or set it globally before pg starts. I tried to set it after
checking the current mode, but it required a restart before taking
affect.