Fixed session cleanup task, for real this time
I didn't read the documentation correctly. The walk function was
skipping everything because the first this it gets is the sessions
directory itself which it then skips. So nothing was being processed.
The directory check will now make sure the path doesn't equal sessions.
For some reason, the counter is not being incremented correctly.
When the walk finishes, w.c never contained how many files it
deleted. That will need to be looked at some more. But, it's at
least working.