Add remote IP to context, return a cached LeaseStore if available
There were several places where the remoteaddr was being parsed
into a new byte slice, it's now part of the context so it doesn't
need to be parsed everytime.

The NewLeaseStore method has been "replace" with the GetLeaseStore
method. It will create a new lease store if needed but otherwise,
return an exising store since it will use the same database in each
instance.