Moved memory using middleware after a route is matched
I believe part of the memory problem is with unregistered devices
resolving every domain name back to the registration server and
when that client makes a request, memory is used for a user and session
object even though they may never be used again. Now, that memory will
only be allocated after a route match is found, meaning the user
deliberatly (or close enough) requested the page. Hopefully this will
cut down the slow increase in memory used.