Removed need to manually import schema, will be done automatically
At applicaiton start, the database is queried for its table list.
If any table is not created, it will be created. The session table
is managed by the database session store and may not be used so
it's not created in this set. First start up may be a bit slower
due to the table creation, but subsequent starts will be just as fast.
This action is one time only on start up and shouldn't impact overall
performance.