Added DHCP lease history
When a lease is saved to the database, it will be sent to a goroutine
that will add it to a lease_history table. If an entry already exists
for the combination of IP, MAC, and start time < now and end time > now,
then the history entry will be updated with a new end time. This will
create a block view for leases where a device may have the same lease
for multiple lease durations. If a lease request ever falls outside
the time window, a new history entry will be created.

The web view hasn't been updated yet to use the system. Once implemented
it will be optional and can be disabled if it causes problems or
erronious results. If it's disabled the current "history" system will
be used instead.