If a single device or username is found, go directly to its page
Previously, if a username or device search returned devices belonging
to the same user or a single device, the search page was still shown
even though the majority of time, an admin would go to the device's or
user's management page. Now, if an exact match is found, the client
is redirected to the respective management page.