Allow authorization via HTTP header
Any /api router can now accept the HTTP basic authentication header
for login. This allows external clients or tools to interact with the
API without having to keep track of a session cookie.

There are two new permissions and configuration settings. The
permissions are APIRead and APIWrite. Although they are mutually
exclusive, the permissions logic will add APIRead if APIWrite is
allowed. By default, no username not even the admin has API permissions.
This is to ensure that allowing API access is intentional.

The new configuration settings are APIReadOnlyUsers and
APIReadWriteUsers. Both are arrays of usernames like the other *Users
settings. By default these lists are empty.