Consistant CLI flags, added -td flag, added man pages
Added the -td flag on the dhcp binary to check a pool configuration
file for syntax errors. Previously, the dhcp binary couldn't check
the syntax of the main configuration file. It could only check
the pool file. Now the -t flag does the same on both binaries.

Created a manual page for each binary to better explain how the
binaries are to be executed.