Embed static assets into binary, allow overrides with custom directory
All templates and public files are now embeded in the built binary
using go-bindata. The template loader has been modified to use it,
and the router is setup with a small package that wraps the assets
and exposes them as an HTTP file system.

Users may override any file by placing them under the same directory
structure under a "custom" directory. E.g. if a user wishes to replace
the logo.png file, they would place their image file at
$programRoot/custom/public/assets/images/logo.png. This file will be
loaded
instead of the embeded one. The configuration files are not embeded and
so a minimal configuration file is still needed.

To build, "make bindata" must be ran. Or simply "make".