Added a status API to check database
The API makes a direct request to the database
for its schema version to ensure the connection
is still active. This can be used with monitoring
software to ensure the database connection doesn't
break again.

This endpoint can be expanded to include memory
information and other status stuff.

Users must have the ViewDebugInfo permission which
is given to all full admins and any one in the
APIStatusUsers group.