Closes #17 - Manage user permissions via the web interface
The old permission settings in the toml file will be accepted,
but ignored. They're used to migrate to the new database
driven system but after that they're not used.

UI and API rights are separate as they were conceptually with
the toml file. All users are in the "default" ui group and the
"disable" api group unless set otherwise. The status API is still
behind a flag as it requires the view debug permission.