Added openid path to initiate OpenID authentication
PG with use the autorization code flow where the user
is redirected to an OpenID login provider and then
back to PG where the access token will be validated
and the username will be retrieved.

No testing has been done yet, and the access token
is not yet verified, not is there any profile information
yet.