Delegates: DB table, user store, registration
Database tables is created to match delegates with their user
and permissions. The user store retrieves and saves delegate information
when the user is saved. The registration handler checks for delegate
permissions with tests to verify implementation.

Still need to integrate with UI to add/edit/delete delegates and for
delegates to switch between accounts.